refactor(chat): centralize tool output preview limits
