+++ frontend/src/ui/chat/tool-calls/renderers/BashCall.tsx
@@ -2,7 +2,8 @@ import { TerminalIcon } from "../../../primitives/icons";
 import type { ToolCallProps } from "../ToolCallTypes";
 import { CodeBlock } from "../CodeBlock";
 import { ToolShell } from "../ToolShell";
-import { DEFAULT_TOOL_OUTPUT_PREVIEW_CHARS, truncate } from "../utils";
+import { DEFAULT_TOOL_OUTPUT_PREVIEW_CHARS } from "../../../../config/chat";
+import { truncate } from "../utils";
 
 export function BashCall({ input, output, outputExpanded, status, isError }: Omit<ToolCallProps, "name">) {
   const command = (input?.command as string) ?? "";
